pyproject.toml
[project]
name = "nano-trm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.5.0",
    "pyyaml>=6.0.2",
    "torch>=2.10.0",
]
